Added Method 3

+++ install.sh
@@ -10,13 +10,21 @@ case $1 in
 		cp all-ways-egpu "${DESTDIR}"${BINDIR}
 		chmod +x "${DESTDIR}"${BINDIR}/all-ways-egpu
 		mkdir -p "${DESTDIR}"${CONFDIR}
-		cp all-ways-egpu.service "${DESTDIR}"${CONFDIR}
-		cp all-ways-egpu-user.service "${DESTDIR}"${CONFDIR}
-		cp all-ways-egpu-boot-vga.service "${DESTDIR}"${CONFDIR}
-		cp all-ways-egpu-shutdown.service "${DESTDIR}"${CONFDIR}
-		cp all-ways-egpu-openrc "${DESTDIR}"${CONFDIR}
-		cp all-ways-egpu-boot-vga-openrc "${DESTDIR}"${CONFDIR}
-		cp all-ways-egpu-reenable.desktop "${DESTDIR}"${CONFDIR}
+		# check if using systemd or openrc
+		if [ -e "$(command -v systemctl)" ]; then
+			cp systemd/all-ways-egpu.service "${DESTDIR}"${CONFDIR}
+			cp systemd/all-ways-egpu-user.service "${DESTDIR}"${CONFDIR}
+			cp systemd/all-ways-egpu-boot-vga.service "${DESTDIR}"${CONFDIR}
+			cp systemd/all-ways-egpu-shutdown.service "${DESTDIR}"${CONFDIR}
+			cp systemd/all-ways-egpu-set-compositor.service "${DESTDIR}"${CONFDIR}
+			cp all-ways-egpu-reenable.desktop "${DESTDIR}"${CONFDIR}
+		else
+			if [ -e "$(command -v rc-status)" ]; then
+				cp OpenRC/all-ways-egpu-openrc "${DESTDIR}"${CONFDIR}
+				cp OpenRC/all-ways-egpu-boot-vga-openrc "${DESTDIR}"${CONFDIR}
+				cp OpenRC/all-ways-egpu-set-compositor-openrc "${DESTDIR}"${CONFDIR}
+			fi
+		fi
 		cp all-ways-egpu.desktop "${DESTDIR}"/usr/share/applications
 		;;
 
@@ -29,7 +37,6 @@ case $1 in
 			cp all-ways-egpu.desktop "$HD"/.local/share/applications
 			echo 'export PATH="$HOME/bin:$PATH"' >> "$HD"/.bashrc
 		done
-#		./.bashrc
 		;;
 
 	uninstall)
